Refactor SIMD to allow non-x86 implementations
The SIMD related code is refactored/relocated to allow a non-x86
SIMD implementation without rewriting the core AoS/SoA conversion
functions.
