+++ .cargo/config.toml
@@ -1,4 +1,4 @@
-[build]
+[target.x86_64-unknown-linux-gnu]
 rustflags = ["-C", "target-cpu=x86-64-v3"]
 
 [env]
